+++ pkg/vm/engine/tae/catalog/catalog.go
@@ -19,7 +19,6 @@ import (
 	"context"
 	"fmt"
 	"sync"
-	"time"
 
 	pkgcatalog "github.com/matrixorigin/matrixone/pkg/catalog"
 	"github.com/matrixorigin/matrixone/pkg/common/moerr"
@@ -177,63 +176,6 @@ func (catalog *Catalog) GCByTS(ctx context.Context, ts types.TS) {
 		panic(err)
 	}
 	catalog.gcTS = ts
-}
-func (catalog *Catalog) CheckMetadata() {
-	logutil.Infof("[MetadataCheck] Start")
-	p := &LoopProcessor{}
-	p.ObjectFn = catalog.checkObject
-	p.TombstoneFn = catalog.checkTombstone
-	catalog.RecurLoop(p)
-}
-func (catalog *Catalog) checkTombstone(t data.Tombstone) error {
-	obj := t.GetObject().(*ObjectEntry)
-	_, err := obj.GetTable().GetObjectByID(&obj.ID)
-	if err != nil {
-		logutil.Infof("[MetadataCheck] tombstone and object doesn't match, err %v, obj %v, tombstone %v",
-			err,
-			obj.PPString(3, 0, ""),
-			t.StringLocked(3, 0, ""))
-	}
-	return nil
-}
-func (catalog *Catalog) checkObject(o *ObjectEntry) error {
-	o.RLock()
-	defer o.RUnlock()
-	if o.Depth() > 2 {
-		logutil.Infof("[MetadataCheck] object mvcc link is too long, depth %d, obj %v", o.Depth(), o.PPStringLocked(3, 0, ""))
-	}
-	if o.IsAppendable() && o.HasDropCommittedLocked() {
-		if o.GetLatestNodeLocked().BaseNode.IsEmpty() {
-			logutil.Infof("[MetadataCheck] object should have stats, obj %v", o.PPStringLocked(3, 0, ""))
-		}
-	}
-	if !o.IsAppendable() && !o.IsCreatingOrAborted() {
-		if o.GetLatestNodeLocked().BaseNode.IsEmpty() {
-			logutil.Infof("[MetadataCheck] object should have stats, obj %v", o.PPStringLocked(3, 0, ""))
-		}
-	}
-	if !o.IsAppendable() && !o.IsCreatingOrAborted() {
-		if o.GetLatestNodeLocked().BaseNode.IsEmpty() {
-			logutil.Infof("[MetadataCheck] object should have stats, obj %v", o.PPStringLocked(3, 0, ""))
-		}
-	}
-	if !catalog.gcTS.IsEmpty() {
-		if o.HasDropCommittedLocked() && o.DeleteBeforeLocked(catalog.gcTS) && !o.InMemoryDeletesExistedLocked() {
-			logutil.Infof("[MetadataCheck] object should not exist, gcTS %v, obj %v", catalog.gcTS.ToString(), o.PPStringLocked(3, 0, ""))
-		}
-	}
-
-	duration := time.Minute * 10
-	ts := types.BuildTS(time.Now().UTC().UnixNano()-duration.Nanoseconds(), 0)
-	if o.HasDropCommittedLocked() && o.DeleteBeforeLocked(ts) {
-		if o.InMemoryDeletesExistedLocked() {
-			logutil.Infof("[MetadataCheck] object has in memory deletes %v after deleted, obj %v, tombstone %v",
-				duration,
-				o.PPStringLocked(3, 0, ""),
-				o.GetTable().TryGetTombstone(o.ID).StringLocked(3, 0, ""))
-		}
-	}
-	return nil
 }
 
 func (catalog *Catalog) Close() error {
